Put plugin unit tests in their own modules
Rather than a generic test_search module, name test modules after the
plugin they're testing. At some stage we should move tests into the
plugins but this will make it easier to do so.
